Use BigInteger for storing ids
Resolve #68

+++ app/models/group.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, String
 from sqlalchemy.orm import relationship
 
 from app.database.base_model import Base
@@ -8,7 +8,7 @@ from app.models.relationships import group_members
 class Group(Base):
     __tablename__ = 'groups'
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String, nullable=False)
     locale = Column(String, nullable=True)
