+++ app/models/request.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, ForeignKey, Integer, String, DateTime
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, String
 from sqlalchemy.orm import relationship
 
 from app.database.base_model import Base
@@ -8,11 +8,11 @@ from app.database.base_model import Base
 class Request(Base):
     __tablename__ = 'requests'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    message_id = Column(Integer, nullable=False)
-    chat_id = Column(Integer, nullable=False)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    message_id = Column(BigInteger, nullable=False)
+    chat_id = Column(BigInteger, nullable=False)
     created_at = Column(DateTime, nullable=False, default=datetime.now)
-    author_id = Column(Integer, ForeignKey('users.id', ondelete='CASCADE', onupdate='CASCADE'), nullable=False)
+    author_id = Column(BigInteger, ForeignKey('users.id', ondelete='CASCADE', onupdate='CASCADE'), nullable=False)
     title = Column(String, nullable=False)
 
     author = relationship("User", back_populates="requests")
